requirements.txt
Adafruit-Blinka==6.13.0
adafruit-circuitpython-ads1x15==2.2.8
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-register==1.9.5
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
gps==3.19
numpy==1.21.2
paho-mqtt==1.5.1
pkg-resources==0.0.0
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
